Leverage ConfigRepository in KafkaApis
