+++ core/src/main/scala/kafka/server/ZkAdminManager.scala
@@ -17,7 +17,7 @@
 package kafka.server
 
 import java.util
-import java.util.{Collections, Properties}
+import java.util.Properties
 
 import kafka.admin.{AdminOperationException, AdminUtils}
 import kafka.common.TopicAlreadyMarkedForDeletionException
@@ -25,22 +25,21 @@ import kafka.log.LogConfig
 import kafka.utils.Log4jController
 import kafka.metrics.KafkaMetricsGroup
 import kafka.server.DynamicConfig.QuotaConfigs
+import kafka.server.metadata.ZkConfigRepository
 import kafka.utils._
 import kafka.utils.Implicits._
 import kafka.zk.{AdminZkClient, KafkaZkClient}
 import org.apache.kafka.clients.admin.{AlterConfigOp, ScramMechanism}
 import org.apache.kafka.clients.admin.AlterConfigOp.OpType
 import org.apache.kafka.common.config.ConfigDef.ConfigKey
-import org.apache.kafka.common.config.{AbstractConfig, ConfigDef, ConfigException, ConfigResource, LogLevelConfig}
+import org.apache.kafka.common.config.{ConfigDef, ConfigException, ConfigResource, LogLevelConfig}
 import org.apache.kafka.common.errors.ThrottlingQuotaExceededException
 import org.apache.kafka.common.errors.{ApiException, InvalidConfigurationException, InvalidPartitionsException, InvalidReplicaAssignmentException, InvalidRequestException, ReassignmentInProgressException, TopicExistsException, UnknownTopicOrPartitionException, UnsupportedVersionException}
-import org.apache.kafka.common.internals.Topic
 import org.apache.kafka.common.message.AlterUserScramCredentialsResponseData.AlterUserScramCredentialsResult
 import org.apache.kafka.common.message.CreatePartitionsRequestData.CreatePartitionsTopic
 import org.apache.kafka.common.message.CreateTopicsRequestData.CreatableTopic
 import org.apache.kafka.common.message.CreateTopicsResponseData.{CreatableTopicConfigs, CreatableTopicResult}
-import org.apache.kafka.common.message.{AlterUserScramCredentialsRequestData, AlterUserScramCredentialsResponseData, DescribeConfigsResponseData, DescribeUserScramCredentialsResponseData}
-import org.apache.kafka.common.message.DescribeConfigsRequestData.DescribeConfigsResource
+import org.apache.kafka.common.message.{AlterUserScramCredentialsRequestData, AlterUserScramCredentialsResponseData, DescribeUserScramCredentialsResponseData}
 import org.apache.kafka.common.message.DescribeUserScramCredentialsResponseData.CredentialInfo
 import org.apache.kafka.common.metrics.Metrics
 import org.apache.kafka.common.security.scram.internals.{ScramMechanism => InternalScramMechanism}
@@ -49,8 +48,7 @@ import org.apache.kafka.server.policy.CreateTopicPolicy.RequestMetadata
 import org.apache.kafka.common.protocol.Errors
 import org.apache.kafka.common.quota.{ClientQuotaAlteration, ClientQuotaEntity, ClientQuotaFilter, ClientQuotaFilterComponent}
 import org.apache.kafka.common.requests.CreateTopicsRequest._
-import org.apache.kafka.common.requests.DescribeConfigsResponse.ConfigSource
-import org.apache.kafka.common.requests.{AlterConfigsRequest, ApiError, DescribeConfigsResponse}
+import org.apache.kafka.common.requests.{AlterConfigsRequest, ApiError}
 import org.apache.kafka.common.security.scram.internals.{ScramCredentialUtils, ScramFormatter}
 import org.apache.kafka.common.utils.Sanitizer
 
@@ -66,6 +64,7 @@ class ZkAdminManager(val config: KafkaConfig,
 
   private val topicPurgatory = DelayedOperationPurgatory[DelayedOperation]("topic", config.brokerId)
   private val adminZkClient = new AdminZkClient(zkClient)
+  private val configHelper = new ConfigHelper(metadataCache, config, new ZkConfigRepository(adminZkClient))
 
   private val createTopicPolicy =
     Option(config.getConfiguredInstance(KafkaConfig.CreateTopicPolicyClassNameProp, classOf[CreateTopicPolicy]))
@@ -116,8 +115,8 @@ class ZkAdminManager(val config: KafkaConfig,
                                               configs: Properties,
                                               assignments: Map[Int, Seq[Int]]): Unit = {
     metadataAndConfigs.get(topicName).foreach { result =>
-      val logConfig = LogConfig.fromProps(LogConfig.extractLogConfigMap(config), configs)
-      val createEntry = createTopicConfigEntry(logConfig, configs, includeSynonyms = false, includeDocumentation = false)(_, _)
+      val logConfig = LogConfig.fromProps(Server.copyKafkaConfigToLog(config), configs)
+      val createEntry = configHelper.createTopicConfigEntry(logConfig, configs, includeSynonyms = false, includeDocumentation = false)(_, _)
       val topicConfigs = logConfig.values.asScala.map { case (k, v) =>
         val entry = createEntry(k, v)
         new CreatableTopicConfigs()
@@ -380,85 +379,6 @@ class ZkAdminManager(val config: KafkaConfig,
     }
   }
 
-  def describeConfigs(resourceToConfigNames: List[DescribeConfigsResource],
-                      includeSynonyms: Boolean,
-                      includeDocumentation: Boolean): List[DescribeConfigsResponseData.DescribeConfigsResult] = {
-    resourceToConfigNames.map { case resource =>
-
-      def allConfigs(config: AbstractConfig) = {
-        config.originals.asScala.filter(_._2 != null) ++ config.nonInternalValues.asScala
-      }
-      def createResponseConfig(configs: Map[String, Any],
-                               createConfigEntry: (String, Any) => DescribeConfigsResponseData.DescribeConfigsResourceResult): DescribeConfigsResponseData.DescribeConfigsResult = {
-        val filteredConfigPairs = if (resource.configurationKeys == null)
-          configs.toBuffer
-        else
-          configs.filter { case (configName, _) =>
-            resource.configurationKeys.asScala.forall(_.contains(configName))
-          }.toBuffer
-
-        val configEntries = filteredConfigPairs.map { case (name, value) => createConfigEntry(name, value) }
-        new DescribeConfigsResponseData.DescribeConfigsResult().setErrorCode(Errors.NONE.code)
-          .setConfigs(configEntries.asJava)
-      }
-
-      try {
-        val configResult = ConfigResource.Type.forId(resource.resourceType) match {
-          case ConfigResource.Type.TOPIC =>
-            val topic = resource.resourceName
-            Topic.validate(topic)
-            if (metadataCache.contains(topic)) {
-              // Consider optimizing this by caching the configs or retrieving them from the `Log` when possible
-              val topicProps = adminZkClient.fetchEntityConfig(ConfigType.Topic, topic)
-              val logConfig = LogConfig.fromProps(LogConfig.extractLogConfigMap(config), topicProps)
-              createResponseConfig(allConfigs(logConfig), createTopicConfigEntry(logConfig, topicProps, includeSynonyms, includeDocumentation))
-            } else {
-              new DescribeConfigsResponseData.DescribeConfigsResult().setErrorCode(Errors.UNKNOWN_TOPIC_OR_PARTITION.code)
-                .setConfigs(Collections.emptyList[DescribeConfigsResponseData.DescribeConfigsResourceResult])
-            }
-
-          case ConfigResource.Type.BROKER =>
-            if (resource.resourceName == null || resource.resourceName.isEmpty)
-              createResponseConfig(config.dynamicConfig.currentDynamicDefaultConfigs,
-                  createBrokerConfigEntry(perBrokerConfig = false, includeSynonyms, includeDocumentation))
-            else if (resourceNameToBrokerId(resource.resourceName) == config.brokerId)
-              createResponseConfig(allConfigs(config),
-                  createBrokerConfigEntry(perBrokerConfig = true, includeSynonyms, includeDocumentation))
-            else
-              throw new InvalidRequestException(s"Unexpected broker id, expected ${config.brokerId} or empty string, but received ${resource.resourceName}")
-
-          case ConfigResource.Type.BROKER_LOGGER =>
-            if (resource.resourceName == null || resource.resourceName.isEmpty)
-              throw new InvalidRequestException("Broker id must not be empty")
-            else if (resourceNameToBrokerId(resource.resourceName) != config.brokerId)
-              throw new InvalidRequestException(s"Unexpected broker id, expected ${config.brokerId} but received ${resource.resourceName}")
-            else
-              createResponseConfig(Log4jController.loggers,
-                (name, value) => new DescribeConfigsResponseData.DescribeConfigsResourceResult().setName(name)
-                  .setValue(value.toString).setConfigSource(ConfigSource.DYNAMIC_BROKER_LOGGER_CONFIG.id)
-                  .setIsSensitive(false).setReadOnly(false).setSynonyms(List.empty.asJava))
-          case resourceType => throw new InvalidRequestException(s"Unsupported resource type: $resourceType")
-        }
-        configResult.setResourceName(resource.resourceName).setResourceType(resource.resourceType)
-      } catch {
-        case e: Throwable =>
-          // Log client errors at a lower level than unexpected exceptions
-          val message = s"Error processing describe configs request for resource $resource"
-          if (e.isInstanceOf[ApiException])
-            info(message, e)
-          else
-            error(message, e)
-          val err = ApiError.fromThrowable(e)
-          new DescribeConfigsResponseData.DescribeConfigsResult()
-              .setResourceName(resource.resourceName)
-              .setResourceType(resource.resourceType)
-              .setErrorMessage(err.message)
-              .setErrorCode(err.error.code)
-              .setConfigs(Collections.emptyList[DescribeConfigsResponseData.DescribeConfigsResourceResult])
-      }
-    }.toList
-  }
-
   def alterConfigs(configs: Map[ConfigResource, AlterConfigsRequest.Config], validateOnly: Boolean): Map[ConfigResource, ApiError] = {
     configs.map { case (resource, config) =>
 
@@ -717,98 +637,6 @@ class ZkAdminManager(val config: KafkaConfig,
       case _: NumberFormatException =>
         throw new InvalidRequestException(s"Broker id must be an integer, but it is: $resourceName")
     }
-  }
-
-  private def brokerSynonyms(name: String): List[String] = {
-    DynamicBrokerConfig.brokerConfigSynonyms(name, matchListenerOverride = true)
-  }
-
-  private def brokerDocumentation(name: String): String = {
-    config.documentationOf(name)
-  }
-
-  private def configResponseType(configType: Option[ConfigDef.Type]): DescribeConfigsResponse.ConfigType = {
-    if (configType.isEmpty)
-      DescribeConfigsResponse.ConfigType.UNKNOWN
-    else configType.get match {
-      case ConfigDef.Type.BOOLEAN => DescribeConfigsResponse.ConfigType.BOOLEAN
-      case ConfigDef.Type.STRING => DescribeConfigsResponse.ConfigType.STRING
-      case ConfigDef.Type.INT => DescribeConfigsResponse.ConfigType.INT
-      case ConfigDef.Type.SHORT => DescribeConfigsResponse.ConfigType.SHORT
-      case ConfigDef.Type.LONG => DescribeConfigsResponse.ConfigType.LONG
-      case ConfigDef.Type.DOUBLE => DescribeConfigsResponse.ConfigType.DOUBLE
-      case ConfigDef.Type.LIST => DescribeConfigsResponse.ConfigType.LIST
-      case ConfigDef.Type.CLASS => DescribeConfigsResponse.ConfigType.CLASS
-      case ConfigDef.Type.PASSWORD => DescribeConfigsResponse.ConfigType.PASSWORD
-      case _ => DescribeConfigsResponse.ConfigType.UNKNOWN
-    }
-  }
-
-  private def configSynonyms(name: String, synonyms: List[String], isSensitive: Boolean): List[DescribeConfigsResponseData.DescribeConfigsSynonym] = {
-    val dynamicConfig = config.dynamicConfig
-    val allSynonyms = mutable.Buffer[DescribeConfigsResponseData.DescribeConfigsSynonym]()
-
-    def maybeAddSynonym(map: Map[String, String], source: ConfigSource)(name: String): Unit = {
-      map.get(name).map { value =>
-        val configValue = if (isSensitive) null else value
-        allSynonyms += new DescribeConfigsResponseData.DescribeConfigsSynonym().setName(name).setValue(configValue).setSource(source.id)
-      }
-    }
-
-    synonyms.foreach(maybeAddSynonym(dynamicConfig.currentDynamicBrokerConfigs, ConfigSource.DYNAMIC_BROKER_CONFIG))
-    synonyms.foreach(maybeAddSynonym(dynamicConfig.currentDynamicDefaultConfigs, ConfigSource.DYNAMIC_DEFAULT_BROKER_CONFIG))
-    synonyms.foreach(maybeAddSynonym(dynamicConfig.staticBrokerConfigs, ConfigSource.STATIC_BROKER_CONFIG))
-    synonyms.foreach(maybeAddSynonym(dynamicConfig.staticDefaultConfigs, ConfigSource.DEFAULT_CONFIG))
-    allSynonyms.dropWhile(s => s.name != name).toList // e.g. drop listener overrides when describing base config
-  }
-
-  private def createTopicConfigEntry(logConfig: LogConfig, topicProps: Properties, includeSynonyms: Boolean, includeDocumentation: Boolean)
-                                    (name: String, value: Any): DescribeConfigsResponseData.DescribeConfigsResourceResult = {
-    val configEntryType = LogConfig.configType(name)
-    val isSensitive = KafkaConfig.maybeSensitive(configEntryType)
-    val valueAsString = if (isSensitive) null else ConfigDef.convertToString(value, configEntryType.orNull)
-    val allSynonyms = {
-      val list = LogConfig.TopicConfigSynonyms.get(name)
-        .map(s => configSynonyms(s, brokerSynonyms(s), isSensitive))
-        .getOrElse(List.empty)
-      if (!topicProps.containsKey(name))
-        list
-      else
-        new DescribeConfigsResponseData.DescribeConfigsSynonym().setName(name).setValue(valueAsString)
-          .setSource(ConfigSource.TOPIC_CONFIG.id) +: list
-    }
-    val source = if (allSynonyms.isEmpty) ConfigSource.DEFAULT_CONFIG.id else allSynonyms.head.source
-    val synonyms = if (!includeSynonyms) List.empty else allSynonyms
-    val dataType = configResponseType(configEntryType)
-    val configDocumentation = if (includeDocumentation) logConfig.documentationOf(name) else null
-    new DescribeConfigsResponseData.DescribeConfigsResourceResult()
-      .setName(name).setValue(valueAsString).setConfigSource(source)
-      .setIsSensitive(isSensitive).setReadOnly(false).setSynonyms(synonyms.asJava)
-      .setDocumentation(configDocumentation).setConfigType(dataType.id)
-  }
-
-  private def createBrokerConfigEntry(perBrokerConfig: Boolean, includeSynonyms: Boolean, includeDocumentation: Boolean)
-                                     (name: String, value: Any): DescribeConfigsResponseData.DescribeConfigsResourceResult = {
-    val allNames = brokerSynonyms(name)
-    val configEntryType = KafkaConfig.configType(name)
-    val isSensitive = KafkaConfig.maybeSensitive(configEntryType)
-    val valueAsString = if (isSensitive)
-      null
-    else value match {
-      case v: String => v
-      case _ => ConfigDef.convertToString(value, configEntryType.orNull)
-    }
-    val allSynonyms = configSynonyms(name, allNames, isSensitive)
-        .filter(perBrokerConfig || _.source == ConfigSource.DYNAMIC_DEFAULT_BROKER_CONFIG.id)
-    val synonyms = if (!includeSynonyms) List.empty else allSynonyms
-    val source = if (allSynonyms.isEmpty) ConfigSource.DEFAULT_CONFIG.id else allSynonyms.head.source
-    val readOnly = !DynamicBrokerConfig.AllDynamicConfigs.contains(name)
-
-    val dataType = configResponseType(configEntryType)
-    val configDocumentation = if (includeDocumentation) brokerDocumentation(name) else null
-    new DescribeConfigsResponseData.DescribeConfigsResourceResult().setName(name).setValue(valueAsString).setConfigSource(source)
-      .setIsSensitive(isSensitive).setReadOnly(readOnly).setSynonyms(synonyms.asJava)
-      .setDocumentation(configDocumentation).setConfigType(dataType.id)
   }
 
   private def sanitizeEntityName(entityName: String): String =
